Refactored utils and rpc logic, splitted it in 2 distinct modules and added a progress sample
