+++ app/actions/bakingController.ts
@@ -1,19 +1,19 @@
 import { Dispatch } from 'redux';
-import { KeysType, BakingControllerStartOptions } from '../utils/padaria/types';
-import bakingController from '../utils/padaria/bakingController';
+import bakingController, { BakingControllerStartOptions } from '../utils/padaria/bakingController';
+import { KeysType } from '../utils/padaria/types';
 
 export enum ControllerActionTypes {
-  START = 'BAKING_CONTROLLER_START',
-  STOP = 'BAKING_CONTROLLER_STOP'
+    START = 'BAKING_CONTROLLER_START',
+    STOP = 'BAKING_CONTROLLER_STOP'
 }
 
 interface StartAction {
-  type: ControllerActionTypes.START;
-  options ?: BakingControllerStartOptions;
+    type: ControllerActionTypes.START;
+    options ?: BakingControllerStartOptions;
 }
 
 interface StopAction {
-  type: ControllerActionTypes.STOP;
+    type: ControllerActionTypes.STOP;
 }
 
 export type ControllerActions = StartAction | StopAction;
@@ -22,19 +22,19 @@ export type StartControllerPrototype = (keys: KeysType, options:BakingController
 export type StopControllerPrototype = () => void;
 
 const startController:StartControllerPrototype = (keys, options) => {
-  bakingController.start(keys, options);
+    bakingController.start(keys, options);
 
-  return (dispatch:Dispatch) => {
-    dispatch({ type: ControllerActionTypes.START, options });
-  }
+    return (dispatch:Dispatch) => {
+      dispatch({ type: ControllerActionTypes.START, options });
+    }
 }
 
 const stopController:StartControllerPrototype = () => {
-  bakingController.stop();
+    bakingController.stop();
 
-  return (dispatch:Dispatch) => {
-    dispatch({ type: ControllerActionTypes.STOP });
-  }
+    return (dispatch:Dispatch) => {
+      dispatch({ type: ControllerActionTypes.STOP });
+    }
 }
 
 export default { startController, stopController };
