+++ app/constants/info.js
@@ -1,4 +1,4 @@
 export default {
   name: 'Padaria',
-  version: '0.7.0'
+  version: '0.8.0'
 }
