+++ app/reducers/bakingController.ts
@@ -1,23 +1,23 @@
 import { ControllerActionTypes, ControllerActions } from '../actions/bakingController';
-import { BakingControllerState  } from '../utils/padaria/types';
+import { BakingControllerState  } from '../utils/padaria/bakingController';
 
 const defaultState = {
-  active: false,
-  baking: false,
-  endorsing: false,
-  accusing: false
+    active: false,
+    baking: false,
+    endorsing: false,
+    accusing: false
 }
 
 export default (state: BakingControllerState = defaultState, action: ControllerActions) => {
-  switch (action.type) {
-    case ControllerActionTypes.START:
-      return {
-        active: action.options.baking || action.options.endorsing || action.options.accusing,
-        ...action.options
-      };
-      case ControllerActionTypes.STOP:
-      return defaultState;
-    default:
-      return state;
-  }
+    switch (action.type) {
+      case ControllerActionTypes.START:
+        return {
+          active: action.options.baking || action.options.endorsing || action.options.accusing,
+          ...action.options
+        };
+        case ControllerActionTypes.STOP:
+        return defaultState;
+      default:
+        return state;
+    }
 }
