+++ app/utils/padaria/index.js
@@ -4,3 +4,4 @@ export { default as baker } from './baker';
 export { default as crypto } from './crypto';
 export { default as storage } from './storage';
 export { default as rewardController } from './rewardController';
+export { default as rpc } from './rpc';
