+++ app/utils/padaria/utils.d.ts
@@ -1,10 +1,7 @@
-import { number, string } from "prop-types";
-
-export type QueryType = QueryTypes | null;
-
-export type LoadOptions = {
-    nodeAddress: string;
-    apiAddress: string;
+export type WaterMarkType = {
+    blockHeader: Uint8Array;
+    endorsement: Uint8Array;
+    genericOperation: Uint8Array;
 }
 
 export type TezosUnitType = {
@@ -12,25 +9,8 @@ export type TezosUnitType = {
     unit: number;
 }
 
-export type WaterMarkType = {
-    blockHeader: Uint8Array,
-    endorsement: Uint8Array,
-    genericOperation: Uint8Array
-}
-
 export interface UtilsInterface {
-    ready: boolean;
-    nodeAddress: string;
-    apiAddress: string;
-    network: string;
-    networkEpoch: string;
     debug: boolean;
-    cycleLength: number;
-    BLOCKS_PER_COMMITMENT: number;
-    blockTime: number;
-    threshold: number;
-    PowHeader: string;
-    networkConstants: {},
     watermark: WaterMarkType;
     uTEZ:   TezosUnitType;
     mTEZ:   TezosUnitType;
@@ -38,12 +18,6 @@ export interface UtilsInterface {
     KTEZ:   TezosUnitType;
     MTEZ:   TezosUnitType;
     setDebugMode: (mode:boolean) => void,
-    load: (options:LoadOptions) => Promise<boolean>;
-    setCurrentNetwork: () => Promise<void>;
-    setNetworkConstants: () => Promise<void>;
-    getCurrentHead: () => Promise<HeadType>;
-    queryNode: (path:string, type?:QueryType, args:any = {}) => Promise<any>;
-    queryAPI: (path:string, type?:QueryType, args:any = {}) => Promise<any>;
     operationType: (op:{ contents: {kind:string}[] }) => number;
     createProtocolData: (priority:number, powHeader?:string, pow?:string, seed?:string) => string;
     convertUnit: (value:number, to:{char:string, unit:number}, from?:{char:string, unit:number}) => string;
@@ -51,32 +25,6 @@ export interface UtilsInterface {
     getSharePercentage: (balance:number, staking_balance:number) => string;
     getShareReward: (balance:number, staking_balance:number, rewards:number) => number;
     parseTEZWithSymbol: (value:number) => string;
+    firstCycleLevel: (level:number) => number;
+    lastCycleLevel: (level:number) => number;
 }
-
-export interface NetworkConstants {
-    block_reward: string;
-    block_security_deposit: string;
-    blocks_per_commitment: number;
-    blocks_per_cycle: number;
-    blocks_per_roll_snapshot: number;
-    blocks_per_voting_period: number;
-    cost_per_byte: string;
-    endorsement_reward: string;
-    endorsement_security_deposit: string;
-    endorsers_per_block: number;
-    hard_gas_limit_per_block: string;
-    hard_gas_limit_per_operation: string;
-    hard_storage_limit_per_operation: string;
-    max_operation_data_length: number;
-    max_proposals_per_delegate: number;
-    max_revelations_per_block: number;
-    michelson_maximum_type_size: number;
-    nonce_length: number;
-    origination_size: number;
-    preserved_cycles: number;
-    proof_of_work_nonce_size: number;
-    proof_of_work_threshold: string;
-    seed_nonce_revelation_tip: string;
-    time_between_blocks: string[];
-    tokens_per_roll: string;
-};
