+++ package.json
@@ -172,7 +172,7 @@
         "@types/react": "^16.7.18",
         "@types/react-dom": "^16.8.2",
         "@types/react-redux": "^7.0.1",
-        "@types/react-router-dom": "^4.3.1",
+        "@types/react-router-dom": "^5.0.0",
         "@types/react-test-renderer": "^16.8.1",
         "@types/redux-logger": "^3.0.6",
         "@types/sinon": "^7.0.6",
@@ -270,8 +270,8 @@
         "react-dom": "^16.8.3",
         "react-hot-loader": "^4.7.1",
         "react-redux": "^6.0.1",
-        "react-router": "^4.3.1",
-        "react-router-dom": "^4.3.1",
+        "react-router": "^5.0.0",
+        "react-router-dom": "^5.0.0",
         "react-virtualized": "^9.21.0",
         "redux": "^4.0.1",
         "redux-saga": "^1.0.2",
